feat(ui): merge Import and Create into one Add source flow
Import and Create sat side by side in the sources header with nothing to
tell a user which one to press. Both ended at the same server action, so
the split was an artifact of how the two flows were built.

One entry point now offers both ways in on its first step: a URL for
Coral to inspect, or a manifest file to drop or pick. What the user gives
decides the steps that follow, so the step pill starts only once the
branch is known: 2/3 and 3/3 for a URL, 2/2 for a manifest.

- source-add.tsx owns the shared first step, both fetchers, and the
  discard confirmation.
- source-create.tsx keeps the steps after discovery as SourceCreateFlow,
  its draft seeded from what Coral detected.
- source-import.tsx keeps the configure step as SourceImportConfigureForm.
- The import route, its path, and the second header button are gone, and
  both branches report as "Added {name}".
