+++ apps/coral-ui/app/routes.ts
@@ -24,7 +24,6 @@ export default [
       route(routePattern('workspaceSources'), 'routes/sources.tsx', [
         route('discover', 'routes/source-discovery.ts'),
         route('describe', 'routes/source-describe.ts'),
-        route('import', 'routes/source-import.tsx'),
         route('install', 'routes/source-install.tsx'),
         route('oauth-import', 'routes/source-oauth-import.ts'),
         route(':sourceName', 'routes/source-detail.tsx'),
