+++ apps/coral-ui/app/routes/source-install.tsx
@@ -6,7 +6,7 @@ import { runSourcesAction } from './sources-action'
 
 import { requestAuthContext } from '@/auth/server-context'
 import { addToast } from '@/wax/components/toast'
-import { SourceCreateDialog } from '@/views/sources/source-create'
+import { IMPORT_ERROR_TOAST_MS, SourceAddDialog } from '@/views/sources/source-add'
 import { routePath } from '@/routing/routemap'
 import { workspaceFromParams } from '@/lib/workspace-routing'
 
@@ -27,10 +27,20 @@ export async function clientAction({
   serverAction,
 }: Route.ClientActionArgs): Promise<SourcesActionData | Response> {
   const result = await serverAction()
+  if (result?.status === 'error') {
+    // Manifest errors carry parse positions, so they need the longer toast
+    // whether the manifest was written here or handed in whole.
+    addToast('error', {
+      description: result.message,
+      durationMs: IMPORT_ERROR_TOAST_MS,
+      title: result.name ? `Could not add ${result.name}` : 'Could not add that source',
+    })
+    return result
+  }
   if (result?.status !== 'success') return result
 
   addToast('success', {
-    title: `Created ${result.name}`,
+    title: `Added ${result.name}`,
     description: 'The source was validated and installed.',
   })
   return redirect(routePath('workspaceSources', { workspaceId: params.workspaceId }))
@@ -42,13 +52,14 @@ export default function SourceInstallRoute({ actionData, params }: Route.Compone
   const sourcesPath = routePath('workspaceSources', { workspaceId: params.workspaceId })
 
   return (
-    <SourceCreateDialog
+    <SourceAddDialog
       actionData={actionData}
+      describePath={routePath('workspaceSourceDescribe', { workspaceId: params.workspaceId })}
       discoveryPath={routePath('workspaceSourceDiscovery', { workspaceId: params.workspaceId })}
       oauthImportPath={`${sourcesPath}/oauth-import`}
       onOAuthImportComplete={async (name, signal) => {
         addToast('success', {
-          title: `Created ${name}`,
+          title: `Added ${name}`,
           description: 'The source was validated and installed.',
         })
         await revalidator.revalidate()
